// pages-mobile/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)";
  };
  public: {
    Tables: {
// ...
      users: {
        Row: {
          admin: boolean;
          avatar_url: string | null;
          bio: string | null;
          created_at: string | null;
          current_period_end: string | null;
          current_period_start: string | null;
          email: string;
          first_name: string | null;
          id: string;
          last_name: string | null;
          organization_id: string | null;
          phone: string | null;
          plan_expires_at: string | null;
          plan_type: string | null;
          subscription_source: string | null;
          subscription_status: string | null;
          trial_ends_at: string | null;
          trial_started_at: string | null;
          updated_at: string | null;
          user_type: Database["public"]["Enums"]["user_type_enum"] | null;
        };
        Insert: {
          admin?: boolean;
          avatar_url?: string | null;
          bio?: string | null;
          created_at?: string | null;
          current_period_end?: string | null;
          current_period_start?: string | null;
          email: string;
          first_name?: string | null;
          id: string;
          last_name?: string | null;
          organization_id?: string | null;
          phone?: string | null;
          plan_expires_at?: string | null;
          plan_type?: string | null;
          subscription_source?: string | null;
          subscription_status?: string | null;
          trial_ends_at?: string | null;
          trial_started_at?: string | null;
          updated_at?: string | null;
          user_type?: Database["public"]["Enums"]["user_type_enum"] | null;
        };
        Update: {
          admin?: boolean;
          avatar_url?: string | null;
          bio?: string | null;
          created_at?: string | null;
          current_period_end?: string | null;
          current_period_start?: string | null;
          email?: string;
          first_name?: string | null;
          id?: string;
          last_name?: string | null;
          organization_id?: string | null;
          phone?: string | null;
          plan_expires_at?: string | null;
          plan_type?: string | null;
          subscription_source?: string | null;
          subscription_status?: string | null;
          trial_ends_at?: string | null;
          trial_started_at?: string | null;
          updated_at?: string | null;
          user_type?: Database["public"]["Enums"]["user_type_enum"] | null;
        };
        Relationships: [
          {
            foreignKeyName: "users_organization_id_fkey";
            columns: ["organization_id"];
            isOneToOne: false;
            referencedRelation: "organizations";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;